Split utils in and outside of tests
Accidentally was in the wrong place
